Applying changes made for ansible/ansible#35159 over.
